refactor: move storage utils to api
