+++ providers/ic/src/providers/publish/publish.providers.ts
@@ -7,7 +7,7 @@ import {emitDeckPublished, publishDeck} from '../../publish/deck.publish';
 import {emitDocPublished, publishDoc} from '../../publish/doc.publish';
 import {publishDeckMetas, publishDocMetas, updateIndexHtml} from '../../publish/metas.publish';
 import {uploadResources} from '../../publish/resources.publish';
-import {getStorageActor} from '../../utils/storage.utils';
+import {getStorageActor} from '../../api/storage.api';
 import {getIdentity} from '../auth/auth.providers';
 
 export const deckPublish: DeckPublish = async ({
